remove business logic from Subheader component
