+++ src/components/Subheader.js
@@ -2,26 +2,16 @@ import React from "react";
 import { Row, Col } from "react-bootstrap";
 import "bootstrap/dist/css/bootstrap.min.css";
 import { List } from "./List.js";
+import { subheaderHelperFunction } from "../functions/subheaderHelperFunction";
 
 export function Subheader(props) {
   // Defines a component Subheaher with more information about the application.
-  const message = "Select zero to N ";
   const italics = { fontStyle: "italic" };
-  const listItems = [];
-  for (
-    let listItem = 0;
-    listItem < props.dimensionsColumns.length;
-    listItem++
-  ) {
-    listItems.push(
-      <List
-        message={message}
-        columns={props.dimensionsColumns[listItem]}
-        styling={italics}
-        key={listItem}
-      />
-    );
-  }
+  const listItems = subheaderHelperFunction(
+    props.dimensionsColumns,
+    props.dimensionsColumns.length,
+    italics
+  );
 
   return (
     <div>
